pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "model-upgrade-analyzer"
version = "0.1.0"
description = "Analyze Azure AI / LLM model upgrades across code, prompts, configs, and Model IQ data."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Model Upgrade Analyzer" }]
dependencies = [
    "pyyaml>=6.0",
    "openpyxl>=3.1",
    "jinja2>=3.1",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
pptx = ["python-pptx>=0.6.23"]

[project.scripts]
model-upgrade-analyzer = "model_upgrade_analyzer.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["model_upgrade_analyzer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
